+++ llama/sampling_ext.h
@@ -47,7 +47,7 @@ extern "C"
         llama_token id,
         bool apply_grammar);
 
-    int schema_to_grammar(const char *json_schema, char *grammar, int32_t max_len);
+    int schema_to_grammar(const char *json_schema, char *grammar, size_t max_len);
 
 #ifdef __cplusplus
 }
